+++ pydv/pdv.py
@@ -139,6 +139,9 @@ class Command(cmd.Cmd, object):
     xlabel_set_from_curve = True
     ylabel = ''
     ylabel_set_from_curve = True
+    filename_set_from_curve = True
+    record_id = ''
+    record_id_set_from_curve = True
     bordercolor = None
     figcolor = None
     plotcolor = None
@@ -162,6 +165,7 @@ class Command(cmd.Cmd, object):
     gridwidth = 1.0
     showletters = True
     showcurveinlegend = False
+    showrecordidinlegend = False
     xlogscale = False
     ylogscale = False
     titlefont = 'large'
@@ -196,6 +200,11 @@ class Command(cmd.Cmd, object):
     ytickwidth = 1
     yminortickwidth = 0.5
     namewidth = 50
+    xlabelwidth = 10
+    ylabelwidth = 10
+    filenamewidth = 50
+    recordidwidth = 10
+
     updatestyle = False
     linewidth = None
 
@@ -237,6 +246,30 @@ class Command(cmd.Cmd, object):
                 else:
                     self.title = title
                     self.title_set_from_curve = from_curve
+
+    def set_filename(self, filename, from_curve=False):
+        if not from_curve:
+            self.filename = filename
+            self.filename_set_from_curve = from_curve if filename != "" else True
+        else:
+            if self.filename_set_from_curve:
+                if len(self.plotlist) > 1 and filename != self.filename:
+                    self.filename = ''
+                else:
+                    self.filename = filename
+                    self.filename_set_from_curve = from_curve
+
+    def set_record_id(self, record_id, from_curve=False):
+        if not from_curve:
+            self.record_id = record_id
+            self.record_id_set_from_curve = from_curve if record_id != "" else True
+        else:
+            if self.record_id_set_from_curve:
+                if len(self.plotlist) > 1 and record_id != self.title:
+                    self.record_id = ''
+                else:
+                    self.record_id = record_id
+                    self.record_id_set_from_curve = from_curve
 
     ##check for special character/operator commands##
     def precmd(self, line):
@@ -2903,10 +2936,94 @@ class Command(cmd.Cmd, object):
             if self.debug:
                 traceback.print_exc(file=sys.stdout)
     def help_namewidth(self):
-        print('\n   Command: change the width of the first column of the menu and lst output. If no width is given, the'
+        print('\n   Command: change the width of the curve_name column of the menu and lst output. If no width is given, the'
               '\n            current column width will be displayed.'
               '\n   Usage:  namewidth [width]')
 
+    ## adjust the width of the xlabel column in 'menu' and 'lst' commands
+    def do_xlabelwidth(self, line):
+        try:
+            if len(line) == 0:
+                print('xlabel column width is currently', self.xlabelwidth)
+            else:
+                line = line.split()
+                width = int(line[0])
+                if width < 0:
+                    width = 0
+                self.xlabelwidth = width
+                print('changing xlabel column width to', self.xlabelwidth)
+        except:
+            print('error - usage: xlabelwidth [width]')
+            if self.debug:
+                traceback.print_exc(file=sys.stdout)
+    def help_xlabelwidth(self):
+        print('\n   Command: change the width of the xlabel column of the menu and lst output. If no width is given, the'
+              '\n            current column width will be displayed.'
+              '\n   Usage:  xlabelwidth [width]')
+
+    ## adjust the width of the ylabel column in 'menu' and 'lst' commands
+    def do_ylabelwidth(self, line):
+        try:
+            if len(line) == 0:
+                print('label column width is currently', self.ylabelwidth)
+            else:
+                line = line.split()
+                width = int(line[0])
+                if width < 0:
+                    width = 0
+                self.ylabelwidth = width
+                print('changing ylabel column width to', self.ylabelwidth)
+        except:
+            print('error - usage: ylabelwidth [width]')
+            if self.debug:
+                traceback.print_exc(file=sys.stdout)
+    def help_ylabelwidth(self):
+        print('\n   Command: change the width of the ylabel column of the menu and lst output. If no width is given, the'
+              '\n            current column width will be displayed.'
+              '\n   Usage:  ylabelwidth [width]')
+
+
+    ## adjust the width of the file column in 'menu' and 'lst' commands
+    def do_filenamewidth(self, line):
+        try:
+            if len(line) == 0:
+                print('file column width is currently', self.filenamewidth)
+            else:
+                line = line.split()
+                width = int(line[0])
+                if width < 0:
+                    width = 0
+                self.filenamewidth = width
+                print('changing file column width to', self.filenamewidth)
+        except:
+            print('error - usage: filenamewidth [width]')
+            if self.debug:
+                traceback.print_exc(file=sys.stdout)
+    def help_filenamewidth(self):
+        print('\n   Command: change the width of the file column of the menu and lst output. If no width is given, the'
+              '\n            current column width will be displayed.'
+              '\n   Usage:  filenamewidth [width]')
+
+    ## adjust the width of the rec id column in 'menu' and 'lst' commands
+    def do_recordidwidth(self, line):
+        try:
+            if len(line) == 0:
+                print('record_id column width is currently', self.recordidwidth)
+            else:
+                line = line.split()
+                width = int(line[0])
+                if width < 0:
+                    width = 0
+                self.recordidwidth = width
+                print('changing rec id column width to', self.recordidwidth)
+        except:
+            print('error - usage: recordidwidth [width]')
+            if self.debug:
+                traceback.print_exc(file=sys.stdout)
+    def help_recordidwidth(self):
+        print('\n   Command: change the width of the record_id column of the menu and lst output. If no width is given, the'
+              '\n            current column width will be displayed.'
+              '\n   Usage:  recordidwidth [width]')
 
     ##adjust the length of the lines in the legend##
     def do_handlelength(self, line):
@@ -2925,7 +3042,6 @@ class Command(cmd.Cmd, object):
     def help_handlelength(self):
         print('\n   Command: change the length of the lines in the legend')
         print('     Usage: handlelength <integer>')
-
 
     ##show or hide minor ticks##
     def do_minorticks(self, line):
@@ -3238,12 +3354,18 @@ class Command(cmd.Cmd, object):
             self.__mod_curve(line, 'lnstyle')
             self.plotedit = True
         except:
-            print('error - usage: lnstyle <curve-list> <style: solid | dash | dot | dashdot>')
+            print('error - usage: lnstyle <curve-list> <style: solid | dash | dot | dashdot '
+                  '| loosely_dotted | long_dash_with_offset | loosely_dashed | dashed '
+                  '| loosely_dashdotted | dashdotted | densely_dashdotted '
+                  '| dashdotdotted | loosely_dashdotdotted | densely_dashdotdotted>')
             if self.debug:
                 traceback.print_exc(file=sys.stdout)
     def help_lnstyle(self):
         print('\n   Procedure: Set the line style of curves'
-              '\n   Usage: lnstyle <curve-list> <style: solid | dash | dot | dashdot>\n')
+              '\n   Usage: lnstyle <curve-list> <style: solid | dash | dot | dashdot '
+                  '| loosely_dotted | long_dash_with_offset | loosely_dashed | dashed '
+                  '| loosely_dashdotted | dashdotted | densely_dashdotted '
+                  '| dashdotdotted | loosely_dashdotdotted | densely_dashdotdotted>\n')
 
     ##set draw style of given curves##
     def do_drawstyle(self, line):
@@ -3281,6 +3403,82 @@ class Command(cmd.Cmd, object):
         [4, 2, 2, 2, 2, 2] : Gives a dash-dot-dot pattern.
    See matplotlib 'set_dashes' command for more information.
 ''')
+
+    def do_group(self, line):
+
+        pn, cn, fn = self.do_listr('1')
+        
+        # Setting Linestyles at the file level
+        files = []
+
+        for f in fn: # ordered set
+            if f not in files:
+                files.append(f)
+                        
+        groups = {}
+        
+        for f in files:
+            temp = []
+            for i, plotname in enumerate(pn):
+                if fn[i] == f:
+                    temp.append(plotname)
+            
+            groups[f] = temp
+            
+        styles = [
+                    ('solid', 'solid'),      # Same as (0, ()) or '-'
+                    ('dot', 'dot'),    # Same as (0, (1, 1)) or ':'
+                    ('dash', 'dash'),    # Same as '--'
+                    ('dashdot', 'dashdot'),  # Same as '-.'
+                    ('loosely_dotted',        '(0, (1, 10))'),
+                    ('long_dash_with_offset', '(5, (10, 3))'),
+                    ('loosely_dashed',        '(0, (5, 10))'),
+                    ('dashed',                '(0, (5, 5))'),
+                    ('loosely_dashdotted',    '(0, (3, 10, 1, 10))'),
+                    ('dashdotted',            '(0, (3, 5, 1, 5))'),
+                    ('densely_dashdotted',    '(0, (3, 1, 1, 1))'),
+                    ('dashdotdotted',         '(0, (3, 5, 1, 5, 1, 5))'),
+                    ('loosely_dashdotdotted', '(0, (3, 10, 1, 10, 1, 10))'),
+                    ('densely_dashdotdotted', '(0, (3, 1, 1, 1, 1, 1))')
+                 ]
+
+        for i, filename in enumerate(groups):
+            if i<14:
+                curves_ = " ".join(groups[filename])
+                self.do_lnstyle(curves_+' '+styles[i][0].replace("'",""))
+            else:
+                print('There are only fourteen linestyles available. Please reduce the number of files.')
+            
+        # Setting Colors at the curve level
+        curve_names = []
+
+        for curve_name  in cn: # ordered set
+            temp_cn = curve_name.split(' - ')[0]
+            if temp_cn not in curve_names:
+                curve_names.append(temp_cn)
+                
+        groups = {}
+        
+        for curve_name in curve_names:
+            temp = []
+            for i, plotname in enumerate(pn):
+                if cn[i].split(' - ')[0] == curve_name:
+                    temp.append(plotname)
+            
+            groups[curve_name] = temp
+        
+        colors = plt.rcParams['axes.prop_cycle'].by_key()['color']
+
+        for i, curve_name in enumerate(groups):
+            if i<10:
+                curves_ = " ".join(groups[curve_name])
+                self.do_color(curves_+' '+colors[i].replace("'",""))
+            else:
+                print('There are only ten colors available. Please reduce the number of same name curves.')
+
+        self.updateplot
+    def help_group(self):
+        print('\n   Group curves based on name and file if curve names are the same.\n')
 
     ##turn hiding on for given curves##
     def do_hide(self, line):
@@ -3379,6 +3577,12 @@ class Command(cmd.Cmd, object):
                 except:
                     print("error - invalid label-pattern")
                     return 0
+            
+            print("{:<5} {:<{namewidth}.{namewidth}} {:<{xlabelwidth}.{xlabelwidth}} {:<{ylabelwidth}.{ylabelwidth}} {:<9} {:<9} {:<9} {:<9} {:<{filenamewidth}.{filenamewidth}} {:<{recordidwidth}.{recordidwidth}}"
+                  .format('curve', 'curve_name', 'xlabel', 'ylabel', 'xmin', 'xmax', 'ymin', 'ymax', 'fname', 'record_id',
+                          namewidth=self.namewidth, xlabelwidth= self.xlabelwidth, ylabelwidth=self.ylabelwidth, filenamewidth=self.filenamewidth,recordidwidth=self.recordidwidth))
+            print("".join(['-']*(5+self.namewidth+self.xlabelwidth+self.ylabelwidth+9+9+9+9+self.filenamewidth+self.recordidwidth+9)))
+
 
             for curve in self.plotlist:
                 searchline = curve.name + ' ' + curve.filename
@@ -3387,20 +3591,31 @@ class Command(cmd.Cmd, object):
                     if curve.edited:
                         plotname = "*"
                     plotname = plotname + curve.plotname
+                    name = curve.name
                     name = pdvutil.truncate(curve.name.ljust(self.namewidth), self.namewidth)
+                    xlabel = curve.xlabel
+                    xlabel = xlabel.ljust(self.xlabelwidth)
+                    xlabel = pdvutil.truncate(xlabel, self.xlabelwidth)
+                    ylabel = curve.ylabel
+                    ylabel = ylabel.ljust(self.ylabelwidth)
+                    ylabel = pdvutil.truncate(ylabel, self.ylabelwidth)
                     fname = curve.filename
+                    fname = fname.ljust(self.filenamewidth)
+                    fname = pdvutil.truncate(fname, self.filenamewidth,'right')
+                    record_id = curve.record_id
+                    record_id = record_id.ljust(self.recordidwidth)
+                    record_id = pdvutil.truncate(record_id, self.recordidwidth)
                     xmin = "%.2e" % min(curve.x)
                     xmax = "%.2e" % max(curve.x)
                     ymin = "%.2e" % min(curve.y)
                     ymax = "%.2e" % max(curve.y)
-                    print("{:>5} {} {:9} {:9} {:9} {:9} {}".format(plotname, name, xmin, xmax, ymin, ymax, fname))
+                    print("{:>5} {} {} {} {:9} {:9} {:9} {:9} {} {}".format(plotname, name, xlabel, ylabel, xmin, xmax, ymin, ymax, fname, record_id))
         except:
             print("error - usage: list [<label-pattern>]")
             if self.debug:
                 traceback.print_exc(file=sys.stdout)
         finally:
             self.redraw = False
-
     def help_list(self):
         print("\n    {}\n    {}\n    {}\n".format("Macro: Display curves in list",
                                                   "Usage: list [<label-pattern>]",
@@ -3408,6 +3623,11 @@ class Command(cmd.Cmd, object):
 
     ##list currently graphed curves##
     def do_listr(self, line):
+        
+        group_plotnames = []
+        group_curvenames = []
+        group_filenames = []
+
         try:
             if not line:
                 self.help_listr()
@@ -3434,25 +3654,46 @@ class Command(cmd.Cmd, object):
             if stop > pllen:
                 stop = pllen
 
+            print("{:<5} {:<{namewidth}.{namewidth}} {:<{xlabelwidth}.{xlabelwidth}} {:<{ylabelwidth}.{ylabelwidth}} {:<9} {:<9} {:<9} {:<9} {:<{filenamewidth}.{filenamewidth}} {:<{recordidwidth}.{recordidwidth}}"
+                  .format('curve', 'curve_name', 'xlabel', 'ylabel', 'xmin', 'xmax', 'ymin', 'ymax', 'fname', 'record_id', 
+                          namewidth=self.namewidth, xlabelwidth= self.xlabelwidth, ylabelwidth=self.ylabelwidth, filenamewidth=self.filenamewidth,recordidwidth=self.recordidwidth))
+            print("".join(['-']*(5+self.namewidth+self.xlabelwidth+self.ylabelwidth+9+9+9+9+self.filenamewidth+self.recordidwidth+9)))
+
             for i in range(start, stop):
                 curve = self.plotlist[i]
                 plotname = ""
                 if curve.edited:
                     plotname = "*"
                 plotname = plotname + curve.plotname
+                name = curve.name
                 name = pdvutil.truncate(curve.name.ljust(self.namewidth), self.namewidth)
+                xlabel = curve.xlabel
+                xlabel = xlabel.ljust(self.xlabelwidth)
+                xlabel = pdvutil.truncate(xlabel, self.xlabelwidth)
+                ylabel = curve.ylabel
+                ylabel = ylabel.ljust(self.ylabelwidth)
+                ylabel = pdvutil.truncate(ylabel, self.ylabelwidth)
                 fname = curve.filename
+                fname = fname.ljust(self.filenamewidth)
+                fname = pdvutil.truncate(fname, self.filenamewidth,'right')
+                record_id = curve.record_id
+                record_id = record_id.ljust(self.recordidwidth)
+                record_id = pdvutil.truncate(record_id, self.recordidwidth)
                 xmin = "%.2e" % min(curve.x)
                 xmax = "%.2e" % max(curve.x)
                 ymin = "%.2e" % min(curve.y)
                 ymax = "%.2e" % max(curve.y)
-                print("{:>5} {} {:9} {:9} {:9} {:9} {}".format(plotname, name, xmin, xmax, ymin, ymax, fname))
+                print("{:>5} {} {} {} {:9} {:9} {:9} {:9} {} {}".format(plotname, name, xlabel, ylabel, xmin, xmax, ymin, ymax, fname, record_id))
+                group_plotnames.append(plotname)
+                group_curvenames.append(name)
+                group_filenames.append(fname)
         except:
             print("error - usage: listr <start> [stop]")
             if self.debug:
                 traceback.print_exc(file=sys.stdout)
         finally:
             self.redraw = False
+            return group_plotnames, group_curvenames, group_filenames
     def help_listr(self):
         print("\n   Macro: Display curves in range from start to stop in the list. If stop is not specified, it will"
               "\n          be set to the end of the plot list."
@@ -3518,17 +3759,34 @@ class Command(cmd.Cmd, object):
             if stop > len(self.curvelist):
                 stop = len(self.curvelist)
 
+            print("{:>5} {:<{namewidth}.{namewidth}} {:<{xlabelwidth}.{xlabelwidth}} {:<{ylabelwidth}.{ylabelwidth}} {:<9} {:<9} {:<9} {:<9} {:<{filenamewidth}.{filenamewidth}} {:<{recordidwidth}.{recordidwidth}}"
+                  .format('index', 'curve_name', 'xlabel', 'ylabel', 'xmin', 'xmax', 'ymin', 'ymax', 'fname', 'record_id', 
+                          namewidth=self.namewidth, xlabelwidth=self.xlabelwidth, ylabelwidth=self.ylabelwidth, filenamewidth=self.filenamewidth,recordidwidth=self.recordidwidth))
+            print("".join(['-']*(5+self.namewidth+self.xlabelwidth+self.ylabelwidth+9+9+9+9+self.filenamewidth+self.recordidwidth+9)))            
+
+
             for i in range(start, stop):
                 index = str(i + 1)
                 name = self.curvelist[i].name
                 name = name.ljust(self.namewidth)
                 name = pdvutil.truncate(name, self.namewidth)
+                xlabel = self.curvelist[i].xlabel
+                xlabel = xlabel.ljust(self.xlabelwidth)
+                xlabel = pdvutil.truncate(xlabel, self.xlabelwidth)
+                ylabel = self.curvelist[i].ylabel
+                ylabel = ylabel.ljust(self.ylabelwidth)
+                ylabel = pdvutil.truncate(ylabel, self.ylabelwidth)
                 fname = self.curvelist[i].filename
+                fname = fname.ljust(self.filenamewidth)
+                fname = pdvutil.truncate(fname, self.filenamewidth,'right')
+                record_id = self.curvelist[i].record_id
+                record_id = record_id.ljust(self.recordidwidth)
+                record_id = pdvutil.truncate(record_id, self.recordidwidth)
                 xmin = "%.2e" % min(self.curvelist[i].x)
                 xmax = "%.2e" % max(self.curvelist[i].x)
                 ymin = "%.2e" % min(self.curvelist[i].y)
                 ymax = "%.2e" % max(self.curvelist[i].y)
-                print("{:>5} {} {:9} {:9} {:9} {:9} {}".format(index, name, xmin, xmax, ymin, ymax, fname))
+                print("{:>5} {} {} {} {:9} {:9} {:9} {:9} {} {}".format(index, name, xlabel, ylabel, xmin, xmax, ymin, ymax, fname, record_id))
         except:
             print("error - usage: menur <start> [stop]")
             if self.debug:
@@ -3551,6 +3809,11 @@ class Command(cmd.Cmd, object):
                     print("error: invalid expression")
                     return 0
 
+            print("{:>5} {:<{namewidth}.{namewidth}} {:<{xlabelwidth}.{xlabelwidth}} {:<{ylabelwidth}.{ylabelwidth}} {:<9} {:<9} {:<9} {:<9} {:<{filenamewidth}.{filenamewidth}} {:<{recordidwidth}.{recordidwidth}}"
+                  .format('index', 'curve_name', 'xlabel', 'ylabel', 'xmin', 'xmax', 'ymin', 'ymax', 'fname', 'record_id', 
+                          namewidth=self.namewidth, xlabelwidth=self.xlabelwidth, ylabelwidth=self.ylabelwidth, filenamewidth=self.filenamewidth,recordidwidth=self.recordidwidth))
+            print("".join(['-']*(5+self.namewidth+self.xlabelwidth+self.ylabelwidth+9+9+9+9+self.filenamewidth+self.recordidwidth+9)))
+
             for i in range(len(self.curvelist)):
                 searchline = self.curvelist[i].name + ' ' + self.curvelist[i].filename
                 if not line or reg.search(searchline):
@@ -3558,12 +3821,23 @@ class Command(cmd.Cmd, object):
                     name = self.curvelist[i].name
                     name = name.ljust(self.namewidth)
                     name = pdvutil.truncate(name, self.namewidth)
+                    xlabel = self.curvelist[i].xlabel
+                    xlabel = xlabel.ljust(self.xlabelwidth)
+                    xlabel = pdvutil.truncate(xlabel, self.xlabelwidth)
+                    ylabel = self.curvelist[i].ylabel
+                    ylabel = ylabel.ljust(self.ylabelwidth)
+                    ylabel = pdvutil.truncate(ylabel, self.ylabelwidth)
                     fname = self.curvelist[i].filename
+                    fname = fname.ljust(self.filenamewidth)
+                    fname = pdvutil.truncate(fname, self.filenamewidth,'right')
+                    record_id = self.curvelist[i].record_id
+                    record_id = record_id.ljust(self.recordidwidth)
+                    record_id = pdvutil.truncate(record_id, self.recordidwidth)
                     xmin = "%.2e" % min(self.curvelist[i].x)
                     xmax = "%.2e" % max(self.curvelist[i].x)
                     ymin = "%.2e" % min(self.curvelist[i].y)
                     ymax = "%.2e" % max(self.curvelist[i].y)
-                    print("{:>5} {} {:9} {:9} {:9} {:9} {}".format(index, name, xmin, xmax, ymin, ymax, fname))
+                    print("{:>5} {} {} {} {:9} {:9} {:9} {:9} {} {}".format(index, name, xlabel, ylabel, xmin, xmax, ymin, ymax, fname, record_id))
         except:
             print("error - usage: menu [<regex>]")
             if self.debug:
@@ -4187,7 +4461,13 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
         icur1, icur2 = _extract_curvelist_number(arg0), _extract_curvelist_number(arg1)
         xc1, yc1 = numpy.array(self.curvelist[icur1].x), numpy.array(self.curvelist[icur1].y)
         xc2, yc2 = numpy.array(self.curvelist[icur2].x), numpy.array(self.curvelist[icur2].y)
-        nc = curve.Curve('', '%s vs %s' % (arg0, arg1))
+        newfilename = ''
+        newrecord_id = ''
+        if self.curvelist[icur2].filename == self.curvelist[icur1].filename:
+            newfilename = self.curvelist[icur2].filename 
+            if self.curvelist[icur2].record_id == self.curvelist[icur1].record_id:
+                newrecord_id = self.curvelist[icur2].record_id
+        nc = curve.Curve(newfilename, '%s vs %s' % (arg0, arg1), newrecord_id, self.curvelist[icur2].ylabel, self.curvelist[icur1].ylabel)
         nc.x = yc2
         nc.y = numpy.interp(xc2, xc1, yc1)
         self.addtoplot(nc)
@@ -4993,9 +5273,26 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
     def help_label(self):
         print('\n   Procedure: Change the key and list label for a curve\n   Usage: label <curve> <new-label>\n')
 
+    ##change label for a curve to the recordid##
+    def do_labelrecordids(self, line):
+            try:
+                line = line.strip()
+                if line == '0' or line.upper() == 'OFF':
+                    self.showrecordidinlegend = False
+                elif line == '1' or line.upper() == 'ON':
+                    self.showrecordidinlegend = True
+                else:
+                    raise RuntimeError('invalid input: requires on or off as argument')
+            except:
+                print('error - usage: labelrecordids <on | off>')
+                if self.debug:
+                    traceback.print_exc(file=sys.stdout)
+    def help_labelrecordids(self):
+        print('\n   Variable: Add curve recordid to the legend label if "on", otherwise hide curve recordid if "off".'
+              '\n   Usage: labelrecordids <on | off>')
 
     ##change label for a curve to the filename##
-    def do_labelfilenames(self, line):
+    def do_labelfilenames(self):
         try:
             for j in range(len(self.plotlist)):
                 cur = self.plotlist[j]
@@ -5014,7 +5311,7 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
               '\n   Usage: labelfilenames\n')
 
 
-    ##change label for a curve to the filename##
+    ##change label for a curve to the curve letter##
     def do_labelcurve(self, line):
             try:
                 line = line.strip()
@@ -5110,7 +5407,7 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
                     fname = os.getenv('HOME') + fname[1:]
                 f = open(fname, 'r')
                 funcfile = f.read()
-                funcs = re.findall('do_\w+', funcfile)
+                funcs = re.findall(r'do_\w+', funcfile)
                 exec(funcfile)
                 #print locals()
 
@@ -5780,6 +6077,10 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
         self.set_xlabel(cur.xlabel, from_curve=True)
         self.set_ylabel(cur.ylabel, from_curve=True)
         self.set_title(cur.title, from_curve=True)
+        # check interference with labelfilenames cmd
+        self.set_filename(cur.filename, from_curve=True)
+        self.set_record_id(cur.record_id, from_curve=True)
+
 
     ##return derivative of curve##
     def derivative(self, cur):
@@ -5869,6 +6170,26 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
                             cur.linestyle = '--'
                         elif modvalue == 'dashdot':
                             cur.linestyle = '-.'
+                        elif modvalue == 'loosely_dotted':
+                            cur.linestyle = (0, (1, 10))
+                        elif modvalue == 'long_dash_with_offset':
+                            cur.linestyle = (5, (10, 3))
+                        elif modvalue == 'loosely_dashed':
+                            cur.linestyle = (0, (5, 10))
+                        elif modvalue == 'dashed':
+                            cur.linestyle = (0, (5, 5))
+                        elif modvalue == 'loosely_dashdotted':
+                            cur.linestyle = (0, (3, 10, 1, 10))
+                        elif modvalue == 'dashdotted':
+                            cur.linestyle = (0, (3, 5, 1, 5))
+                        elif modvalue == 'densely_dashdotted':
+                            cur.linestyle = (0, (3, 1, 1, 1))
+                        elif modvalue == 'dashdotdotted':
+                            cur.linestyle = (0, (3, 5, 1, 5, 1, 5))
+                        elif modvalue == 'loosely_dashdotdotted':
+                            cur.linestyle = (0, (3, 10, 1, 10, 1, 10))
+                        elif modvalue == 'densely_dashdotdotted':
+                            cur.linestyle = (0, (3, 1, 1, 1, 1, 1))                           
                         cur.dashes = None      # Restore default dash behaviour
                     elif(flag == 'drawstyle'):
                         # default, steps, steps-pre, steps-post
@@ -6349,8 +6670,9 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
                 self.plotter.canvas.draw()
                 return 0
 
-            # Show curve letter in legend is enabled
+            # Show in legend if enabled
             for cur in self.plotlist:
+                # Show curve letter in legend if enabled
                 addstr = str('[' + cur.plotname + ']')
                 if cur.name.find(addstr) != -1:
                     strarr = cur.name.split(addstr)
@@ -6358,6 +6680,15 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
 
                 if self.showcurveinlegend:
                     cur.name = addstr + ' ' + cur.name
+
+                # Show curve recordid in legend if enabled
+                addrstr = str('- ' + cur.record_id)
+                if cur.name.find(addrstr) != -1:
+                    rstrarr = cur.name.split(addrstr)
+                    cur.name = ''.join(rstrarr).strip()
+                if self.showrecordidinlegend:
+                    cur.name =  cur.name + ' ' + addrstr
+
 
             #set scaling and tick locations
             #
@@ -6587,6 +6918,14 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
                         self.title = val
                     elif(var == 'namewidth'):
                         self.namewidth = int(val)
+                    elif(var == 'xlabelwidth'):
+                        self.xlabelwidth = int(val)
+                    elif(var == 'ylabelwidth'):
+                        self.ylabelwidth = int(val)
+                    elif(var == 'filenamewidth'):
+                        self.filenamewidth = int(val)
+                    elif(var == 'recordidwidth'):
+                        self.recordidwidth = int(val)
                     elif(var == 'key'):
                         if(val.upper() == 'ON' or val == str(1)):
                             self.showkey = True
@@ -6678,7 +7017,7 @@ For a painfully complete explanation of the regex syntax, type 'help regex'.
 
     def console_run(self):
         while True:
-            self.cmdloop('\n\tPython Data Visualizer 3.1.5  -  12.12.2022\n\tType "help" for more information.\n\n')
+            self.cmdloop('\n\tPython Data Visualizer 3.1.6  -  04.12.2023\n\tType "help" for more information.\n\n')
             print('\n   Starting Python Console...\n   Ctrl-D to return to PyDV\n')
             console = code.InteractiveConsole(locals())
             console.interact()